Move SharedState to state.rs file
