+++ src/lib.rs
@@ -3,6 +3,7 @@ pub mod async_connection;
 pub mod error;
 pub mod message;
 pub mod requestable;
+pub mod state;
 pub mod util;
 pub mod zero_connection;
